Move terminal client to new location
